fix accessory and fix some get api

+++ StoreManagement/Controllers/Phase2/CheckingController.cs
@@ -16,21 +16,17 @@ namespace StoreManagement.Controllers
 
         // Count
 
-
         [Route("api/ReceivingController/GetPackedIdentifyItemByBoxIDCounted")]
         [HttpGet]
         public IHttpActionResult GetPackedIdentifyItemByBoxIDCounted(string boxID)
         {
             List<IdentifiedItem> identifiedItems;
             List<Client_IdentifiedItem> client_IdentifiedItems = new List<Client_IdentifiedItem>();
+            BoxController boxController = new BoxController();
             try
             {
-                Box box = (from b in db.Boxes
-                           where b.BoxID == boxID
-                           select b).FirstOrDefault();
-                UnstoredBox uBox = (from uB in db.UnstoredBoxes
-                                    where uB.BoxPK == box.BoxPK
-                                    select uB).FirstOrDefault();
+                Box box = boxController.GetBoxByBoxID(boxID);
+                UnstoredBox uBox = boxController.GetUnstoredBoxbyBoxPK(box.BoxPK);
                 identifiedItems = (from iI in db.IdentifiedItems.OrderByDescending(unit => unit.PackedItemPK)
                                    where iI.UnstoredBoxPK == uBox.UnstoredBoxPK && iI.IsCounted == true
                                    select iI).ToList();
@@ -66,13 +62,59 @@ namespace StoreManagement.Controllers
 
             return Content(HttpStatusCode.OK, client_IdentifiedItems);
         }
+        // đợi Tùng fix accessory
+        [Route("api/ReceivingController/GetEmployeeCountingSessions")]
+        [HttpGet]
+        public IHttpActionResult GetEmployeeCountingSessions(string userID)
+        {
+            List<CountingSession> countingSessions = new List<CountingSession>();
+
+            try
+            {
+                //countingSessions = (from cSS in db.CountingSessions
+                //                    where cSS.userID == userID
+                //                    select cSS).ToList();
+
+                //identifiedItems = (from iI in db.IdentifiedItems.OrderByDescending(unit => unit.PackedItemPK)
+                //                   where iI.UnstoredBoxPK == uBox.UnstoredBoxPK && iI.IsCounted == true
+                //                   select iI).ToList();
+
+                //foreach (var identifiedItem in identifiedItems)
+                //{
+                //    PackedItem packedItem = (from pI in db.PackedItems
+                //                             where pI.PackedItemPK == identifiedItem.PackedItemPK
+                //                             select pI).FirstOrDefault();
+                //    // lấy pack ID
+                //    Pack pack = (from p in db.Packs
+                //                 where p.PackPK == packedItem.PackPK
+                //                 select p).FirstOrDefault();
+
+                //    // lấy phụ liệu tương ứng
+                //    OrderedItem orderedItem = (from oI in db.OrderedItems
+                //                               where oI.OrderedItemPK == packedItem.OrderedItemPK
+                //                               select oI).FirstOrDefault();
+
+                //    Accessory accessory = (from a in db.Accessories
+                //                           where a.AccessoryPK == orderedItem.AccessoryPK
+                //                           select a).FirstOrDefault();
+
+                //    client_IdentifiedItems.Add(new Client_IdentifiedItem(identifiedItem, accessory, pack.PackID));
+                //}
+            }
+            catch (Exception e)
+            {
+                return Content(HttpStatusCode.Conflict, new Content_InnerException(e).InnerMessage());
+            }
+
+            return Content(HttpStatusCode.OK, countingSessions);
+        }
 
         [Route("api/ReceivingController/CountItemBusiness")]
         [HttpPost]
-        public IHttpActionResult CountItemBusiness(int identifiedItemPK, int countedQuantity, string employeeCode)
+        public IHttpActionResult CountItemBusiness(int identifiedItemPK, int countedQuantity, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 4)
             {
@@ -99,7 +141,7 @@ namespace StoreManagement.Controllers
                             if (!identifiedItem.IsCounted)
                             {
                                 // tạo session update và iscounted
-                                countingItemController.createCountingSession(new CountingSession(identifiedItemPK, countedQuantity, employeeCode));
+                                countingItemController.createCountingSession(new CountingSession(identifiedItemPK, countedQuantity, userID));
                                 countingItemController.updateIsCountedOfIdentifiedItem(identifiedItemPK, true);
                             }
                             else
@@ -134,10 +176,10 @@ namespace StoreManagement.Controllers
 
         [Route("api/ReceivingController/EditCountItemBusiness")]
         [HttpPut]
-        public IHttpActionResult EditCountItemBusiness(int countingSessionPK, int countedQuantity, string employeeCode)
+        public IHttpActionResult EditCountItemBusiness(int countingSessionPK, int countedQuantity, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 4)
             {
@@ -164,10 +206,10 @@ namespace StoreManagement.Controllers
 
         [Route("api/ReceivingController/DeleteCountItemBusiness")]
         [HttpDelete]
-        public IHttpActionResult DeleteCountItemBusiness(int countingSessionPK, string employeeCode)
+        public IHttpActionResult DeleteCountItemBusiness(int countingSessionPK, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 4)
             {
@@ -198,10 +240,10 @@ namespace StoreManagement.Controllers
         // Check
         [Route("api/ReceivingController/CheckItemBusiness")]
         [HttpPost]
-        public IHttpActionResult CheckItemBusiness(int identifiedItemPK, int checkedQuantity, int unqualifiedQuantity, string employeeCode)
+        public IHttpActionResult CheckItemBusiness(int identifiedItemPK, int checkedQuantity, int unqualifiedQuantity, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 6)
             {
@@ -227,7 +269,7 @@ namespace StoreManagement.Controllers
                             if (!identifiedItem.IsChecked)
                             {
                                 // tạo session update và ischecked
-                                checkingItemController.createCheckingSession(new CheckingSession(checkedQuantity, unqualifiedQuantity, identifiedItemPK, employeeCode));
+                                checkingItemController.createCheckingSession(new CheckingSession(checkedQuantity, unqualifiedQuantity, identifiedItemPK, userID));
                                 checkingItemController.updateIsCheckedOfIdentifiedItem(identifiedItemPK, true);
                             }
 
@@ -259,10 +301,10 @@ namespace StoreManagement.Controllers
 
         [Route("api/ReceivingController/EditCheckItemBusiness")]
         [HttpPut]
-        public IHttpActionResult EditCheckItemBusiness(int checkingSessionPK, int checkedQuantity, int unqualifiedQuantity, string employeeCode)
+        public IHttpActionResult EditCheckItemBusiness(int checkingSessionPK, int checkedQuantity, int unqualifiedQuantity, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 6)
             {
@@ -289,10 +331,10 @@ namespace StoreManagement.Controllers
 
         [Route("api/ReceivingController/DeleteCheckItemBusiness")]
         [HttpDelete]
-        public IHttpActionResult DeleteCheckItemBusiness(int checkingSessionPK, string employeeCode)
+        public IHttpActionResult DeleteCheckItemBusiness(int checkingSessionPK, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 6)
             {
@@ -323,10 +365,10 @@ namespace StoreManagement.Controllers
         // Classify
         [Route("api/ReceivingController/ClassifyItemBusiness")]
         [HttpPost]
-        public IHttpActionResult ClassifyItemBusiness(int packedItemPK, string comment, int qualityState, string employeeCode)
+        public IHttpActionResult ClassifyItemBusiness(int packedItemPK, string comment, int qualityState, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 6)
             {
@@ -350,23 +392,32 @@ namespace StoreManagement.Controllers
                         // nếu có classify item của packitem thì edit
                         if (tempItem != null)
                         {
-                            if (classifyingItemController.isNotStoredOrReturned(tempItem.ClassifiedItemPK))
+                            ClassifyingSession tempSS = (from cS in db.ClassifyingSessions
+                                                         where cS.ClassifiedItemPK == tempItem.ClassifiedItemPK
+                                                         select cS).FirstOrDefault();
+                            if (userID.Equals(tempSS.UserID))
                             {
-                                // tạo failed or passed item
-                                classifyingItemController.manageItemByQualityState(tempItem.ClassifiedItemPK, tempItem.QualityState, qualityState);
-                                // edit
-                                tempItem.QualityState = qualityState;
-                                ClassifyingSession tempSS = (from cS in db.ClassifyingSessions
-                                                             where cS.ClassifiedItemPK == tempItem.ClassifiedItemPK
-                                                             select cS).FirstOrDefault();
-                                tempSS.Comment = comment;
-                                classifyingItemController.updateClassifiedItem(tempItem);
-                                classifyingItemController.updateClassifyingSession(tempSS);
+                                if (classifyingItemController.isNotStoredOrReturned(tempItem.ClassifiedItemPK))
+                                {
+                                    // tạo failed or passed item
+                                    classifyingItemController.manageItemByQualityState(tempItem.ClassifiedItemPK, tempItem.QualityState, qualityState);
+                                    // edit
+                                    tempItem.QualityState = qualityState;
+
+                                    tempSS.Comment = comment;
+                                    classifyingItemController.updateClassifiedItem(tempItem);
+                                    classifyingItemController.updateClassifyingSession(tempSS);
+                                }
+                                else
+                                {
+                                    return Content(HttpStatusCode.Conflict, "ITEM ĐÃ ĐƯỢC TRẢ HOẶC LƯU KHO");
+                                }
                             }
                             else
                             {
-                                return Content(HttpStatusCode.Conflict, "ITEM ĐÃ ĐƯỢC TRẢ HOẶC LƯU KHO");
+                                return Content(HttpStatusCode.Conflict, "UserID KHÔNG HỢP LỆ VÌ KHÔNG PHẢI LÀ NGƯỜI TẠO");
                             }
+
                         }
                         // nếu chưa có classify item của packitem thì tạo mới
                         else
@@ -378,7 +429,7 @@ namespace StoreManagement.Controllers
                             classifiedItem = classifyingItemController.createClassifiedItem(classifiedItem);
 
                             // tạo classifying session
-                            classifyingItemController.createClassifyingSession(new ClassifyingSession(comment, classifiedItem.ClassifiedItemPK, employeeCode));
+                            classifyingItemController.createClassifyingSession(new ClassifyingSession(comment, classifiedItem.ClassifiedItemPK, userID));
 
                             // đổi IsClassified của pack item
                             packedItem.IsClassified = true;
@@ -410,10 +461,10 @@ namespace StoreManagement.Controllers
 
         [Route("api/ReceivingController/DeleteClassifyItemBusiness")]
         [HttpDelete]
-        public IHttpActionResult DeleteClassifyBusiness(int classifyingSessionPK, string employeeCode)
+        public IHttpActionResult DeleteClassifyBusiness(int classifyingSessionPK, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 6)
             {
@@ -425,23 +476,29 @@ namespace StoreManagement.Controllers
                 {
                     // init
                     ClassifyingSession classifyingSession = db.ClassifyingSessions.Find(classifyingSessionPK);
-                    ClassifiedItem classifiedItem = db.ClassifiedItems.Find(classifyingSession.ClassifiedItemPK);
-                    PackedItem packedItem = db.PackedItems.Find(classifiedItem.PackedItemPK);
-
-                    if (classifyingItemController.isNotStoredOrReturned(classifiedItem.ClassifiedItemPK))
+                    if (userID.Equals(classifyingSession.UserID))
                     {
-                        // delete
-                        classifyingItemController.deleteClassifyingSession(classifyingSession.ClassifyingSessionPK);
-                        classifyingItemController.deleteItemByQualityState(classifiedItem.ClassifiedItemPK, classifiedItem.QualityState);
-                        classifyingItemController.deleteClassifiedItem(classifiedItem.ClassifiedItemPK);
-                        packedItem.IsClassified = false;
-                        packedItemsController.isUpdatedPackedItem(packedItem);
+                        ClassifiedItem classifiedItem = db.ClassifiedItems.Find(classifyingSession.ClassifiedItemPK);
+                        PackedItem packedItem = db.PackedItems.Find(classifiedItem.PackedItemPK);
+
+                        if (classifyingItemController.isNotStoredOrReturned(classifiedItem.ClassifiedItemPK))
+                        {
+                            // delete
+                            classifyingItemController.deleteClassifyingSession(classifyingSession.ClassifyingSessionPK);
+                            classifyingItemController.deleteItemByQualityState(classifiedItem.ClassifiedItemPK, classifiedItem.QualityState);
+                            classifyingItemController.deleteClassifiedItem(classifiedItem.ClassifiedItemPK);
+                            packedItem.IsClassified = false;
+                            packedItemsController.isUpdatedPackedItem(packedItem);
+                        }
+                        else
+                        {
+                            return Content(HttpStatusCode.Conflict, "ITEM ĐÃ ĐƯỢC TRẢ HOẶC LƯU KHO");
+                        }
                     }
                     else
                     {
-                        return Content(HttpStatusCode.Conflict, "ITEM ĐÃ ĐƯỢC TRẢ HOẶC LƯU KHO");
-                    }
 
+                    }
                 }
                 catch (Exception e)
                 {
@@ -459,10 +516,10 @@ namespace StoreManagement.Controllers
 
         [Route("api/ReceivingController/ReturnItemBusiness")]
         [HttpPost]
-        public IHttpActionResult ReturnItemBusiness(int failedItemPK, string employeeCode)
+        public IHttpActionResult ReturnItemBusiness(int failedItemPK, string userID)
         {
             // kiểm trước khi chạy lệnh
-            SystemUser systemUser = db.SystemUsers.Find(employeeCode);
+            SystemUser systemUser = db.SystemUsers.Find(userID);
             // check role of system user
             if (systemUser != null && systemUser.RoleID == 6)
             {
@@ -471,7 +528,7 @@ namespace StoreManagement.Controllers
                 // chạy lệnh classify
                 try
                 {
-                    returningItemController.createReturningSession(failedItemPK, employeeCode);
+                    returningItemController.createReturningSession(failedItemPK, userID);
                     returningItemController.updateFailedItemIsReturned(failedItemPK);
                     returningItemController.updateAllIdentifiedItemsToVirtualBox(failedItemPK);
                 }
@@ -488,7 +545,6 @@ namespace StoreManagement.Controllers
             }
 
         }
-
 
     }
 }
